Wrap up Mesh Concept and Gf[] and Gf() rework
- Remove comma operator, use C++23 multiarg operator[]
- Drastically simplify evaluation mechanism
- Evaluation now mesh based evaluate(mesh_t const &m, auto const &func, point_t p)
- Rename linear_index -> datidx, index -> idx in C++ and Python
- Replace range() with range::all
- Remove cluster_mesh, mesh::brzone and mesh::cyclat now standalone
- Generalize [block_]gf expr for multi-arg bracket
- Add MatsubaraFreq type in python, allow in GfImFreq Evaluation
- Add mesh operator[] and to_idx, to_datidx to mesh python bindings
- Extend port_to_triqs3 script
- Use nda GPU branch under review
- [clang-tidy] Disable warnings about magic numbers and virtual class destructors
- [cmake] Use DEV_GFV2 branch from TRIQS/h5
- Use h5 namespaced signature, e.g. 'h5::read' where possible
- Use nodiscard more consistently
- Add any_of concept to limit to a list of types
- In gf_expr::mesh() only call combine_mesh the first time
- Include mesh_point<M> in backward compat aliases
- Remove discrete mesh, merge into mesh/index.hpp and mesh/legendre.hpp
- Add imaginary time evaluation to legendre gf
- Allow MeshPoint in call to Gf
- Move mesh specific py_converters into their own file
- Remove domain from meshes and concept, expose parameters such as 'beta' and 'statistic' directly as members
- For all meshes, add operator()(idx_t) -> mesh_point_t

+++ test/c++/atom_diag/atom_diag_complex.cpp
@@ -353,10 +353,10 @@ TEST(atom_diag_complex, Functions) {
   // GF
   gf_struct_t gf_struct = {{"dn", 3}, {"up", 3}};
 
-  int n_tau   = 1000;
-  size_t n_iw = 400;
-  int n_l     = 25;
-  int n_w     = 1000;
+  int n_tau = 1000;
+  int n_iw  = 400;
+  int n_l   = 25;
+  int n_w   = 1000;
 
   triqs::atom_diag::excluded_states_t excluded_states = {{1, 0}, {1, 1}, {3, 0}, {3, 1}, {3, 2}, {3, 3}};
   auto G_tau                                          = atomic_g_tau(ad, beta, gf_struct, n_tau, excluded_states);
